% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxColBreaks.R
\name{auxColBreaks}
\alias{auxColBreaks}
\title{Calculate color breaks for raster contrast stretch}
\usage{
auxColBreaks(raster)
}
\arguments{
\item{raster}{a raster data file}
}
\value{
vector of values for color break points
}
\description{
The function computes break values for a contrast stretched raster
}
\examples{
Not run:
breaks <- auxColBreaks(raster = raster)
}

