% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxGridTics.R
\name{auxGridTics}
\alias{auxGridTics}
\title{Calculate position of map grid tics}
\usage{
auxGridTics(gdf, gnbr = 5)
}
\arguments{
\item{gdf}{a geo-data file}

\item{gnbr}{number of map grid lines (horizontal = vertical)}
}
\value{
data frame with grid tic positions
}
\description{
The function computes the position of map grid tics
}
\examples{
Not run:
grid <- auxGridTics(gdf = raster, gnbr = 5)
}

