% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/magicVis-package.R
\docType{package}
\name{magicVis-package}
\alias{magicVis-package}
\alias{remote}
\title{Smorgasboard for visualization functions}
\description{
Smorgasboard for visualization functions
}
\details{
The package provides a variety of general to highly specialized visualization
 functions.
}
\author{
Tim Appelhans, Florian Detsch, Hanna Meyer, Thomas Nauss, Insa Otte\cr
\cr
\emph{Maintainer:} Environmental Informatics \email{admin@environmentalinformatics-marburg.de}
}
\keyword{package}

