% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapAttributeOverRaster.R
\name{mapAttributeOverRaster}
\alias{mapAttributeOverRaster}
\title{Visualize point vector attributes on top of a raster layer}
\usage{
mapAttributeOverRaster(raster, vector, vector_colors = "default", gnbr = 5,
  attr = 3)
}
\arguments{
\item{raster}{a raster layer}

\item{vector}{a spatial point data frame}

\item{vector_colors}{a color palette for the vector layer}

\item{gnbr}{number of map grid lines (horizontal = vertical)}

\item{attr}{column number of the attribute to be visualized}
}
\value{
map-type lattice plot object
}
\description{
The function plots a raster map as background for a point vector layer.
}
\examples{
Not run:
map(landsat, vegetation, gnbr=3, attr=4)
}

