% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapVectorSimple.R
\name{mapVectorSimple}
\alias{mapVectorSimple}
\title{Visualize point vector attributes}
\usage{
mapVectorSimple(vector, attr = 1, colors = "default", gnbr = 5,
  glns = FALSE)
}
\arguments{
\item{vector}{a spatial point data frame}

\item{attr}{column number of the attribute to be visualized}

\item{colors}{a color palette for the vector layer}

\item{gnbr}{number of map grid lines (horizontal = vertical)}

\item{glns}{draw grid lines (TRUE) or just tics (FALSE)}
}
\value{
map-type lattice plot object
}
\description{
The function plots a point vector map.
}
\examples{
library(raster)
library(sp)

data(meuse)
coordinates(meuse) = ~x+y
vec <- meuse[5]

mapVectorSimple(vec, attr = 1)
}

